Pull the kernel from cpio archive.
The ramdisk is now a compressed archive of the kernel image,
modules and configuration data wrapped up into a CPIO file
compiled into the loader image.  We decompress it to RAM
immediately before the loader text, and extract and load the
kernel from that.

The kernel contract has also changed to accept the physical
address and length of the archive as arguments to the entry
function (e.g., illumos Oxide arch `_start`).

These changes update `phbl` to reflect this new reality.

Signed-off-by: Dan Cross <cross@oxidecomputer.com>