Pass address of ramdisk to host OS for real.
Previously, `_start()` in illumos had (possibly) taken
a pointer to a boot properties list, and passing in the
physical address of the ramdisk/cpio archive caused an
early panic, so we hard-coded the argument to constant
zero.

Now that `_start()` is being modified to take proper
arguments, pass the ramdisk physical address as intended.

Tested on Ethanol-X and Gimlet.

Signed-off-by: Dan Cross <cross@oxidecomputer.com>