fix cfg guard so that cpio blob is included
The existing guard around including the cpio blob attempts to match on
the target triple, but is using the "target_arch" attribute.  This
attribute does not have the triple as the value, just "x86_64".
Instead, we can use "target_os" and "target_vendor" to check for the
standalone target we use when building the binary for inclusion in the
ROM.