trivial: Clean up information some output
On my most recent testing run, I noticed this output:

Decompressing cpio archive to 0x0000000077358000..0x7f358000...Done.

One may argue that the phrasing here could use improvement,
perhaps rightfully so.  But notice the difference between the
two numbers, and in particular the leading zeroes of the first
and lack thereof in the second.  This is due to debug formatting
of a pointer for the first, and a usize as the second.  While
minor, I found this jarring; this PR cleans that up by
formatting both as usize.

Signed-off-by: Dan Cross <cross@oxidecomputer.com>