KFlags: Pass flags to kernel
Arrange to pass flags to the loaded kernel.  These mirror the
flags we could send via IPCC, and adds a "machine type" which
(nominally) tells the loaded kernel what kind of machine it will
run on.

This allows the kernel to detect that we're running on, e.g., a
dev machine based on a reference design in the same image we
would use on an Oxide machine, where in the latter we use the
IPCC protocol to communicate with a service processor.

We can specify different combinations of flags and machine types
using cargo features.

Tested on sn14.

Signed-off-by: Dan Cross <cross@oxidecomputer.com>