allocator: align arena on 4KiB boundaries
I noticed when moving the BSS around that the page allocator
for building page tables (e.g., for mapping the kernel) was
inheriting a region that was not 4KiB aligned.  The simplest
fix was just to constrain the allocator itself to be aligned
on a 4K boundary, and there was no reason not to do this.

Signed-off-by: Dan Cross <cross@oxidecomputer.com>