phbl.ld: Work around an LLD bug
For some time I have wanted the ability to link phbl with lld,
but the text segment was always oddly sized.  I found that by
adding an empty section in a PT_NULL segment, I can work around
that bug and the program links correctly (and this is backwards
compatible with GNU ld).

Signed-off-by: Dan Cross <cross@oxidecomputer.com>