update: Update rust and dependencies.
Note the change to the platform definition; the ABI requires
some kind of float support (even though I really don't want it),
and I don't want to go through the trouble of turning on the
FPU, so `+soft-float` it is.  Note that this also requires us
to add the `rustc-abi` token to the target config file.

Also bring us up to date with the new edition, and rename
`rust-toolchain` to `rust-toolchain.toml`.

Signed-off-by: Dan Cross <cross@oxidecomputer.com>